Code cleanup

+++ cards-service/src/main/resources/data.sql
@@ -2,9 +2,9 @@ DROP TABLE IF EXISTS cards;
 
 CREATE TABLE `cards`
 (
-    `card_id`          int          NOT NULL AUTO_INCREMENT,
+    `card_id`          bigint       NOT NULL AUTO_INCREMENT,
     `card_number`      varchar(100) NOT NULL,
-    `customer_id`      int          NOT NULL,
+    `customer_id`      bigint       NOT NULL,
     `card_type`        varchar(100) NOT NULL,
     `total_limit`      int          NOT NULL,
     `amount_used`      int          NOT NULL,
